feat: Add export to Excel functionality
Add a new button and modal for exporting data to Excel. This feature allows users to export data from the application to an Excel file. The button is added to the homepage.html file and the modal is added to the main.js file. This enhancement improves the usability of the application by providing a convenient way to export data.